backend/requirements.txt
# FastAPI Backend Requirements

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# CORS
python-cors==1.0.0
